fix(pipeline): close narrative quality review gaps
