+++ client/src/components/pipeline/AutopilotPanel.jsx
@@ -1,8 +1,8 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { Link } from 'react-router';
 import {
-  Rocket, Loader2, X, Sliders, ShieldCheck, AlertCircle, CheckCircle2,
-  PauseCircle, Play, ScanSearch, ChevronRight, ChevronDown,
+  Rocket, Loader2, X, Sliders, AlertCircle, CheckCircle2,
+  PauseCircle, Play, ChevronRight, ChevronDown,
 } from 'lucide-react';
 import toast from '../ui/Toast';
 import { usePipelineProgress } from '../../hooks/usePipelineProgress';
@@ -14,7 +14,6 @@ import {
   getPipelineAutopilotStatus,
   getPipelineAutopilotModelMetrics,
   pipelineAutopilotSseUrl,
-  getPipelineSeriesCanonReadiness,
   getPipelineSeries,
   listPipelineIssues,
   getProviders,
@@ -26,6 +25,7 @@ import { autopilotStepLabel, describeAutopilotVerification, autopilotMarkerTermi
 import Pill from '../ui/Pill';
 import ProviderModelSelector from '../ProviderModelSelector';
 import AutopilotMilestones from './AutopilotMilestones';
+import CanonReadinessPanel from './CanonReadinessPanel';
 import SeriesAutopilotSchedule from './SeriesAutopilotSchedule';
 import { severityColor } from './constants.js';
 
@@ -550,8 +550,6 @@ export default function AutopilotPanel({ series, onSeriesUpdate, onIssuesUpdate 
   // scheduled unattended runs of every series. Always starts unticked, so
   // clearing locks is a fresh, deliberate choice each run.
   const [unlockForRun, setUnlockForRun] = useState(false);
-  const [canon, setCanon] = useState(null);
-  const [canonLoading, setCanonLoading] = useState(false);
   // Per-run provider/model override. '' = "use the series default", i.e. the
   // provider picked in the series header (`series.llm`), falling back to the
   // install's active provider — the same chain the server resolves in
@@ -937,14 +935,6 @@ export default function AutopilotPanel({ series, onSeriesUpdate, onIssuesUpdate 
   const pause = useCallback(async () => {
     const result = await pausePipelineAutopilot(seriesId).catch(() => null);
     setPausePending(result?.pauseRequested === true);
-  }, [seriesId]);
-
-  const checkCanon = useCallback(async () => {
-    setCanonLoading(true);
-    const report = await getPipelineSeriesCanonReadiness(seriesId, { silent: true })
-      .catch((err) => { toast.error(err.message || 'Canon check failed'); return null; });
-    setCanonLoading(false);
-    if (report) setCanon(report);
   }, [seriesId]);
 
   if (!seriesId) return null;
@@ -1504,58 +1494,7 @@ export default function AutopilotPanel({ series, onSeriesUpdate, onIssuesUpdate 
       ) : null}
 
       {/* Production readiness (canon descriptive integrity) */}
-      <div className="px-3 pb-3 border-t border-port-border pt-2">
-        <div className="flex items-center gap-2">
-          <ShieldCheck size={13} className="text-gray-400" />
-          <span className="text-xs text-gray-300">Production readiness — are all drawn characters/places/objects described?</span>
-          <button
-            type="button"
-            onClick={checkCanon}
-            disabled={canonLoading}
-            className="ml-auto inline-flex items-center gap-1 px-2 py-1 rounded text-xs text-gray-300 hover:text-white border border-port-border bg-port-bg hover:border-port-accent/40 disabled:opacity-40"
-          >
-            {canonLoading ? <Loader2 size={12} className="animate-spin" /> : <ScanSearch size={12} />}
-            Check
-          </button>
-        </div>
-        {canon ? (
-          canon.ready ? (
-            <p className="mt-2 text-xs text-port-success flex items-center gap-1.5">
-              <CheckCircle2 size={12} /> Every noun that gets drawn has a description.
-            </p>
-          ) : (
-            <div className="mt-2 space-y-2">
-              {(canon.blockingIssues || []).some((bi) => bi.blockingReason === 'missing-visual-source') ? (
-                <p className="text-xs text-port-warning">
-                  Some issues have no finished visual script yet. An outline or prose draft cannot prove canon is ready for rendering.
-                </p>
-              ) : null}
-              {(canon.undescribed || []).length > 0 ? (
-                <p className="text-xs text-port-warning">
-                  {canon.undescribed.length} noun(s) appear where they&apos;d be drawn but have no description — fix before generating art:
-                </p>
-              ) : null}
-              {(canon.blockingIssues || []).map((bi) => (
-                <div key={bi.issueId} className="text-xs">
-                  <Link
-                    to={bi.blockingReason === 'missing-visual-source'
-                      ? `/pipeline/issues/${bi.issueId}/${bi.missingSourceStages?.[0] || 'idea'}`
-                      : `/pipeline/issues/${bi.issueId}/nouns`}
-                    className="text-port-accent hover:underline"
-                  >
-                    #{bi.number} {bi.title || ''} →
-                  </Link>
-                  <span className="text-gray-400">
-                    {' '}{bi.blockingReason === 'missing-visual-source'
-                      ? `missing ${(bi.missingSourceStages || []).join(' + ')}`
-                      : (bi.none || []).map((n) => `${n.name} (${n.kind})`).join(', ')}
-                  </span>
-                </div>
-              ))}
-            </div>
-          )
-        ) : null}
-      </div>
+      <CanonReadinessPanel seriesId={seriesId} />
     </div>
   );
 }
