+++ server/services/pipeline/arcPlanner/volumeVerifyPromptContract.test.js
@@ -74,6 +74,8 @@ const SCOPED_CITATIONS = [
   ['arcPosition', 'issue leaf', issueLeafFields],
   ['arcRole', 'issue leaf', issueLeafFields],
   ['lengthProfile', 'issue leaf', issueLeafFields],
+  ['pageTarget', 'issue leaf', issueLeafFields],
+  ['minutesTarget', 'issue leaf', issueLeafFields],
   // Checks #1/#4/#5 read these off the volume node.
   ['logline', 'volume node', volumeNodeFields],
   ['synopsis', 'volume node', volumeNodeFields],
@@ -160,18 +162,33 @@ describe('renderVolumeIssue depth selection', () => {
   it('renders beats (and never synopsis) once the expand pass has run', () => {
     expect(renderVolumeIssue(issue)).toEqual({
       number: 4, title: 'Example Issue', status: 'draft', arcPosition: 4,
-      arcRole: null, lengthProfile: null, beats: 'beat sheet',
+      arcRole: null, lengthProfile: null, pageTarget: 22, minutesTarget: 24,
+      beats: 'beat sheet',
     });
   });
 
   it('renders synopsis (and never beats) when synopsisOnly is requested', () => {
     expect(renderVolumeIssue(issue, { synopsisOnly: true })).toEqual({
       number: 4, title: 'Example Issue', status: 'draft', arcPosition: 4,
-      arcRole: null, lengthProfile: null, synopsis: 'seed synopsis',
+      arcRole: null, lengthProfile: null, pageTarget: 22, minutesTarget: 24,
+      synopsis: 'seed synopsis',
     });
   });
 
   it('nulls synopsis rather than emitting an empty string for an unseeded issue', () => {
     expect(renderVolumeIssue({ ...issue, stages: {} }).synopsis).toBeNull();
+  });
+
+  it('materializes custom numeric targets for metadata-fit checks', () => {
+    expect(renderVolumeIssue({
+      ...issue,
+      lengthProfile: 'custom',
+      pageTarget: 4,
+      minutesTarget: 120,
+    })).toMatchObject({
+      lengthProfile: 'custom',
+      pageTarget: 4,
+      minutesTarget: 120,
+    });
   });
 });
