+++ server/services/pipeline/editorial/checkRunner.js
@@ -15,23 +15,30 @@
  * late-connecting clients via lib/sseUtils.js.
  */
 
-import { randomUUID, createHash } from 'crypto';
+import { randomUUID } from 'crypto';
 import { createSseRunner } from '../../../lib/sseUtils.js';
 import { runStagedLLM, runInlineLLM, runStageScopedInlineLLM, resolveStageContext } from '../../../lib/stageRunner.js';
 import { planManuscriptPass, fitContextToManuscriptFloor, estimateTokens, MANUSCRIPT_FLOOR_TOKENS } from '../../../lib/contextBudget.js';
-import { getEnabledChecks, getEnabledCheckRows, getAllChecks, applySeriesCheckConfig, orderChecksByDependencies, buildCustomCheck, CUSTOM_CHECK_ID_PREFIX, EDITORIAL_SOURCES, comicLetteringIssues, proseStageIssues } from '../../../lib/editorial/index.js';
+import { getEnabledChecks, getEnabledCheckRows, getAllChecks, applySeriesCheckConfig, orderChecksByDependencies, buildCustomCheck, CUSTOM_CHECK_ID_PREFIX } from '../../../lib/editorial/index.js';
 import { getSettings } from '../../settings.js';
 import { getSeries } from '../series.js';
 import { listIssuesForSeries } from '../issues.js';
 import { getSeriesCanon } from '../seriesCanon.js';
-import { collectManuscriptSections, sectionsCorpus, manuscriptSectionHeader, manuscriptSourceHash } from '../arcPlanner.js';
+import { collectManuscriptSections, sectionsCorpus, manuscriptSectionHeader } from '../arcPlanner.js';
 import { getReverseOutline } from '../reverseOutline.js';
 import { getFactsLedger } from '../continuityBible.js';
 import { getSeriesEditorial } from '../editorialAnalysis.js';
-import { seedReviewFromFindings, getReview } from '../manuscriptReview.js';
+import { seedReviewFromFindings } from '../manuscriptReview.js';
 import { recordTrendSnapshot } from '../editorialScore.js';
 import { readReadinessGate, mergeSeverityWeights } from '../../../lib/editorial/index.js';
-import { canonicalStringify } from '../../../lib/objects.js';
+import {
+  buildEditorialSourceProjection,
+  checkSources,
+  effectiveCheckSources,
+  fingerprintForCheck,
+} from './reviewStaleness.js';
+
+export { effectiveCheckSources, getReviewWithStaleness } from './reviewStaleness.js';
 
 // Per-check severity breakdown for telemetry (#1578). Bucket a check's findings
 // by severity so the autopilot SSE stream can show a per-check high/medium/low
@@ -46,308 +53,6 @@ function severityBreakdown(findings) {
     counts[sev] += 1;
   }
   return counts;
-}
-
-// Source-content fingerprinting for finding staleness (#1345, #1387). Each finding
-// is stamped with a hash of the exact content its check analyzed; the manuscript
-// editor / triage view flags a finding `stale` once that content drifts.
-//
-// Per-check declared sources (#1387): a check declares the inputs its run() reads
-// via `check.sources` (a subset of EDITORIAL_SOURCES), and we fingerprint EXACTLY
-// those — so a naming finding (sources: ['canon']) doesn't go stale on a prose or
-// style-guide edit, and editing the ticking clock stales only the
-// arc.ticking-clock-hygiene finding (sources: ['series.arc.tickingClock']) instead
-// of every canon-only finding. This replaces the prior two-segment heuristic
-// (manuscript-vs-canon) that over-flagged because it folded the style guide +
-// ticking clock into shared segments.
-//
-// `SOURCE_RESOLVERS` maps each declared token to the exact content hashed.
-// `canonicalStringify` (key-sorted) keeps the hash stable across machines so a
-// synced finding isn't falsely flagged stale after an import re-orders keys. A
-// load-time guard asserts every EDITORIAL_SOURCES token has a resolver here — a
-// token with no resolver would silently contribute nothing (false-fresh).
-const HASH_SEP = '\u0000';
-const sha256 = (text) => createHash('sha256').update(text || '').digest('hex');
-const SOURCE_RESOLVERS = {
-  manuscript: ({ manuscript }) => manuscript || '',
-  canon: ({ canon }) => canonicalStringify(canon ?? null),
-  // The continuity-bible facts ledger the timeline / canon-contradiction check
-  // reconciles the prose against (#1581). Fingerprint the whole facts array so a
-  // re-extraction or a hand-edit to any established fact stales its findings.
-  continuityBible: ({ continuityBible }) => canonicalStringify(continuityBible ?? null),
-  'series.styleGuide': ({ series }) => canonicalStringify(series?.styleGuide ?? null),
-  'series.arc.tickingClock': ({ series }) => canonicalStringify(series?.arc?.tickingClock ?? null),
-  // The authored reader-map hooks/payoffs the Chekhov check reconciles against (#1299).
-  'series.arc.readerMap': ({ series }) => canonicalStringify(series?.arc?.readerMap ?? null),
-  // The authored foreshadowing ledger the Chekhov check reconciles against (#2172).
-  // Fingerprint the whole ledger so editing a plant/reinforce/payoff seed stales its findings.
-  'series.arc.foreshadowing': ({ series }) => canonicalStringify(series?.arc?.foreshadowing ?? null),
-  // The authored arc themes the theme.coherence check reconciles the prose against
-  // (#1317). Lives on the already-loaded series record, so no extra I/O — fingerprint
-  // the whole themes array so adding/editing a declared theme stales the findings.
-  'series.arc.themes': ({ series }) => canonicalStringify(series?.arc?.themes ?? null),
-  // The author-supplied real-world fact reference the research.fact-accuracy check
-  // reconciles the prose against (#1588). Lives on the already-loaded series record,
-  // so no extra I/O — fingerprint the reference text so editing it stales fact findings.
-  'series.factReference': ({ series }) => canonicalStringify(series?.factReference ?? null),
-  // The reverse-outline scenes the check reads (#1296). Fingerprinting the whole
-  // scenes array is intentionally over-eager (any scene edit stales a finding)
-  // rather than under: safe vs. false-fresh, and the check reads several scene fields.
-  reverseOutline: ({ reverseOutline }) => canonicalStringify(reverseOutline ?? null),
-  // The reverse-outline PLOTLINES the plot-structure check reconciles dropped
-  // subplots against (#1310). Separate token from `reverseOutline` so a scene
-  // edit that doesn't touch the plotline list doesn't needlessly stale a
-  // plotline-only finding (and vice-versa) — same over-eager-but-safe policy.
-  'reverseOutline.plotlines': ({ reverseOutlinePlotlines }) =>
-    canonicalStringify(reverseOutlinePlotlines ?? null),
-  // The detected per-character arc directions a POV check reads (#1295). The
-  // injected `editorialArcs` is the stable projection (name/arcDirection/issueCount/
-  // isProtagonist) — NOT the raw getSeriesEditorial output, which carries a
-  // per-call `generatedAt` timestamp that would re-stale every finding each run.
-  // The `complete` flag is folded in too: a prose edit that stales the analysis
-  // (without re-running it) leaves the projection byte-identical but flips
-  // completeness, and pov.justified's "absent from arcs" finding depends on that
-  // flag — so a finding must go stale when it changes, not only when the arcs do.
-  editorialArcs: ({ editorialArcs, editorialArcsComplete }) =>
-    canonicalStringify({ arcs: editorialArcs ?? null, complete: editorialArcsComplete === true }),
-  // The AUTHORED per-character story arcs the arc.transitions check reconciles
-  // against (#1293). Lives on the already-loaded series record, so no extra I/O
-  // — fingerprint the whole array so any arc/transition edit stales the findings.
-  'series.characterArcs': ({ series }) => canonicalStringify(series?.characterArcs ?? null),
-  // The per-issue storyboard shot lists the visual.shot-continuity check reads
-  // (#1315). Fingerprint ONLY the fields the check actually reads (scene
-  // heading/slugline + each shot's grammar fields) via `projectStoryboardContinuity`
-  // — NOT the whole scene object, so an unrelated render/status edit
-  // (`imageJobId`, `sceneVideoJobId`, wardrobe metadata) doesn't falsely stale a
-  // continuity finding. Mirrors `projectComicLetteringContent` for the comic check.
-  'storyboard.shots': ({ storyboardScenes }) =>
-    canonicalStringify(projectStoryboardContinuity(storyboardScenes) ?? null),
-  // Every issue's AUTHORITATIVE comic lettering content, keyed by issue number
-  // (#1313). The lettering-density check reads the edited comic-pages split (or the
-  // generated script when unsplit) — NOT the prose manuscript — so it gets its own
-  // source token: editing a comic script/page stales lettering findings without
-  // staling prose findings, and vice-versa. `projectComicLetteringContent` builds
-  // the stable [{ number, panels: [{ caption, dialogue, sfx }] }] off the SAME
-  // `comicLetteringIssues` the check analyzes, so a finding stales exactly when the
-  // text it read changes (and an unrelated image render — `panel.imageJobId` — does
-  // NOT stale it, since only the lettering fields are projected).
-  comicScript: ({ comicScripts }) => canonicalStringify(comicScripts ?? null),
-  // The page-turn check's content (#1314) — its own token because it reads each
-  // panel's visual `description` (+ caption/dialogue/SFX text) for the LLM digest.
-  // A description edit stales a page-turn finding without staling a lettering one
-  // (which doesn't read `description`), and vice-versa.
-  'comicScript.pacing': ({ comicPacingContent }) => canonicalStringify(comicPacingContent ?? null),
-  // The panel-rhythm check's content (#1314) — LAYOUT ONLY: it reads nothing but
-  // the per-page panel COUNT (splash/crowding/grid-monotony verdicts), so its
-  // fingerprint is just the counts. A text-only edit (rewording a caption or
-  // description without adding/removing a panel) must NOT stale a rhythm finding —
-  // the verdict cannot have changed. Distinct from `comicScript.pacing` (which
-  // hashes the text the page-turn LLM reads).
-  'comicScript.layout': ({ comicLayoutContent }) => canonicalStringify(comicLayoutContent ?? null),
-  // Each issue's PROSE-stage text the comic↔prose-sync check compares against the
-  // comic (#1589). Its own token (NOT `manuscript`): the stitched manuscript picks
-  // comicScript over prose for a hybrid issue, so a `manuscript` fingerprint would
-  // track comic edits, not prose. `proseContent` is built off the SAME
-  // `proseStageIssues` projection the check reads (per-issue prose keyed by number),
-  // so a finding stales exactly when the prose it compared changes — and a comic-only
-  // edit doesn't stale it.
-  prose: ({ proseContent }) => canonicalStringify(proseContent ?? null),
-};
-
-// Flatten the storyboard scenes across every issue into the `{ issueNumber, scene }`
-// list the visual.shot-continuity check reads (#1315). Built off the already-loaded
-// issues — no extra I/O. Only issues that actually have storyboard scenes contribute,
-// so a series with no visual stage yields an empty list (the check's gate then skips).
-function collectStoryboardScenes(issues) {
-  const out = [];
-  for (const issue of (Array.isArray(issues) ? issues : [])) {
-    const scenes = issue?.stages?.storyboards?.scenes;
-    if (!Array.isArray(scenes) || !scenes.length) continue;
-    const issueNumber = Number.isInteger(issue.number) ? issue.number : null;
-    for (const scene of scenes) {
-      if (scene && typeof scene === 'object') out.push({ issueNumber, scene });
-    }
-  }
-  return out;
-}
-
-// Project the collected storyboard scenes down to ONLY the fields the
-// visual.shot-continuity check reads (#1315), for the staleness fingerprint —
-// the scene's heading/slugline (its finding location) and each shot's grammar
-// fields (`id`, `continuityFromShotId`, `screenDirection`, `shotType`,
-// `description` — the anchorQuote source). Excludes render/status fields
-// (`imageJobId`, `sceneVideoJobId`, wardrobe, …) so a finding stales only when
-// the shot grammar it analyzed changes, not on an unrelated render. Mirrors
-// `projectComicLetteringContent`. Type-guarded throughout (scenes ride peer sync).
-function projectStoryboardContinuity(storyboardScenes) {
-  return (Array.isArray(storyboardScenes) ? storyboardScenes : []).map(({ issueNumber, scene }) => ({
-    issueNumber: Number.isInteger(issueNumber) ? issueNumber : null,
-    heading: typeof scene?.heading === 'string' ? scene.heading : '',
-    slugline: typeof scene?.slugline === 'string' ? scene.slugline : '',
-    shots: (Array.isArray(scene?.shots) ? scene.shots : []).map((s) => ({
-      id: typeof s?.id === 'string' ? s.id : '',
-      continuityFromShotId: typeof s?.continuityFromShotId === 'string' ? s.continuityFromShotId : null,
-      screenDirection: typeof s?.screenDirection === 'string' ? s.screenDirection : null,
-      shotType: typeof s?.shotType === 'string' ? s.shotType : null,
-      description: typeof s?.description === 'string' ? s.description : '',
-    })),
-  }));
-}
-
-// The three comic projections below all derive from the SAME parsed page list
-// (`comicLetteringIssues(issues)` → `[{ number, pages }]`), so the caller parses
-// the comic scripts ONCE (`comicIssuesFor(issues)`) and passes the rows in —
-// rather than each projection re-parsing every issue's script. They take the
-// already-parsed `comicIssues` rows, not raw `issues`.
-
-// Lettering token (`comicScript`, #1313): keeps ONLY caption/dialogue/SFX — the
-// fields `panelLetteringMetrics` consumes — so the hash is stable across image
-// renders and description edits that don't change lettering. PAGE GROUPING is
-// preserved: the check reports per-page totals/locations, so moving panels between
-// pages must change the hash even when the lettering text is identical.
-function projectComicLetteringContent(comicIssues) {
-  return comicIssues.map(({ number, pages }) => ({
-    number,
-    pages: pages.map((p) => ({
-      panels: (Array.isArray(p?.panels) ? p.panels : []).map((panel) => ({
-        caption: typeof panel?.caption === 'string' ? panel.caption : '',
-        dialogue: Array.isArray(panel?.dialogue) ? panel.dialogue : [],
-        sfx: typeof panel?.sfx === 'string' ? panel.sfx : '',
-      })),
-    })),
-  }));
-}
-
-// Page-turn token (`comicScript.pacing`, #1314): distinct from the lettering token
-// so the two checks' fingerprints don't bleed — editing a panel's visual
-// `description` must stale a page-turn finding (the LLM digest reads it) WITHOUT
-// staling a lettering finding (which never reads it), and vice-versa. So this adds
-// `description` on top of caption/dialogue/SFX. PAGE GROUPING preserved; render/
-// status fields (`panel.imageJobId`) are never projected.
-function projectComicPacingContent(comicIssues) {
-  return comicIssues.map(({ number, pages }) => ({
-    number,
-    pages: pages.map((p) => ({
-      panels: (Array.isArray(p?.panels) ? p.panels : []).map((panel) => ({
-        description: typeof panel?.description === 'string' ? panel.description : '',
-        caption: typeof panel?.caption === 'string' ? panel.caption : '',
-        dialogue: Array.isArray(panel?.dialogue) ? panel.dialogue : [],
-        sfx: typeof panel?.sfx === 'string' ? panel.sfx : '',
-      })),
-    })),
-  }));
-}
-
-// Layout token (`comicScript.layout`, #1314): LAYOUT ONLY — the per-page panel
-// COUNT — for the panel-rhythm check, which reads nothing but counts
-// (`analyzePanelRhythm`). Fingerprinting only the count means a text-only edit
-// (reword a caption/description without changing how many panels a page has) does
-// NOT stale a rhythm finding, while adding/removing/moving a panel does. Per-page
-// array order is preserved so a reordering that changes the run structure (splash
-// runs, monotony) re-hashes.
-function projectComicLayoutContent(comicIssues) {
-  return comicIssues.map(({ number, pages }) => ({
-    number,
-    panelCounts: pages.map((p) => (Array.isArray(p?.panels) ? p.panels.length : 0)),
-  }));
-}
-
-// Stable projection of the series editorial aggregate down to the arc fields a
-// POV/arc check reads — drops the volatile `generatedAt` (and the rest) so the
-// staleness fingerprint only moves when a character's detected arc actually does.
-function projectEditorialArcs(editorial) {
-  const chars = Array.isArray(editorial?.characters) ? editorial.characters : [];
-  return chars.map((c) => ({
-    name: c?.name || '',
-    arcDirection: c?.arcDirection || 'flat',
-    issueCount: Number.isFinite(c?.issueCount) ? c.issueCount : 0,
-    isProtagonist: c?.isProtagonist === true,
-  }));
-}
-
-// True only when every analyzable issue has a fresh, complete analysis — the
-// signal pov.justified uses to tell "absent because arc-less" from "absent
-// because not-yet-analyzed" (#1295). Injected into ctx for the check AND folded
-// into the editorialArcs fingerprint above so a prose edit that stales coverage
-// (without changing the arc projection) still stales the POV findings.
-function editorialCoverageComplete(editorial) {
-  const cov = editorial?.coverage;
-  return !!cov && cov.withContent > 0 && cov.analyzed >= cov.withContent && (cov.stale || 0) === 0;
-}
-for (const token of EDITORIAL_SOURCES) {
-  if (typeof SOURCE_RESOLVERS[token] !== 'function') {
-    throw new Error(`checkRunner: editorial source "${token}" has no fingerprint resolver — keep SOURCE_RESOLVERS in sync with EDITORIAL_SOURCES`);
-  }
-}
-
-// A check's declared sources, falling back to the legacy needsManuscript heuristic
-// for any check synthesized before the declaration existed (e.g. an older custom
-// check). Unknown tokens are dropped so a typo can't corrupt the hash.
-function checkSources(check) {
-  const declared = Array.isArray(check?.sources) && check.sources.length
-    ? check.sources
-    : (check?.needsManuscript ? ['manuscript', 'canon'] : ['canon']);
-  return declared.filter((token) => SOURCE_RESOLVERS[token]);
-}
-
-// The source tokens a check's findings EFFECTIVELY depend on for staleness: its
-// OWN declared sources PLUS, transitively, the sources of every check it lists in
-// `dependsOn` (#1627). A dependency-consuming check reads its dependency's findings
-// via ctx.priorFindings, so its compound finding can change when the DEPENDENCY's
-// source content drifts even though the dependent's own sources didn't. Folding the
-// dependencies' sources in keeps getReviewWithStaleness from marking such a finding
-// falsely fresh — and the matching I/O gates fetch that content so it isn't falsely
-// stale either. Resolved against the full registry (`checkById`) and cycle-guarded
-// so it's identical at seed-time and read-time. With no checkById (or no dependsOn)
-// it returns the check's own sources unchanged, so a check that opts out is untouched.
-export function effectiveCheckSources(check, checkById) {
-  const own = checkSources(check);
-  const deps = Array.isArray(check?.dependsOn) ? check.dependsOn : [];
-  if (!deps.length || !(checkById instanceof Map)) return own;
-  const tokens = new Set(own);
-  const seen = new Set([check?.id]);
-  const stack = [...deps];
-  while (stack.length) {
-    const id = stack.pop();
-    if (seen.has(id)) continue;
-    seen.add(id);
-    const dep = checkById.get(id);
-    if (!dep) continue;
-    for (const token of checkSources(dep)) tokens.add(token);
-    for (const next of (Array.isArray(dep.dependsOn) ? dep.dependsOn : [])) {
-      if (!seen.has(next)) stack.push(next);
-    }
-  }
-  return [...tokens];
-}
-
-// Resolve every source token's content ONCE for a given inputs object
-// (`{ manuscript, canon, series }`), so fingerprinting many checks/comments doesn't
-// re-stringify the canon per call. Returns a token→string map the fingerprint reads.
-function resolveSources(inputs) {
-  const resolved = {};
-  for (const token of EDITORIAL_SOURCES) resolved[token] = SOURCE_RESOLVERS[token](inputs);
-  return resolved;
-}
-
-// Fingerprint exactly the inputs a check reads, from a pre-resolved token→content
-// map (see `resolveSources`). Tokens are de-duped and sorted so the hash is
-// independent of declaration order; each segment is prefixed with its token so two
-// source sets can't collide on equal content. NUL joins segments (it can't appear
-// in the JSON the resolvers emit) so they can't run together ambiguously.
-function fingerprintForCheck(check, resolved, checkById = null) {
-  const segments = [...new Set(effectiveCheckSources(check, checkById))]
-    .sort()
-    .map((token) => `${token}=${resolved[token]}`);
-  // A custom check's run logic IS its authored prompt (user data, not code), so a
-  // prompt edit must stale its prior findings even when the manuscript is unchanged
-  // — fold it into the fingerprint. Built-in checks' logic lives in code (a code
-  // change isn't user content and isn't fingerprinted), so only their declared
-  // content sources matter. (#1346, #1387)
-  if (check?.isCustom && typeof check.prompt === 'string') {
-    segments.push(`definition=${check.prompt}`);
-  }
-  return sha256(segments.join(HASH_SEP));
 }
 
 // Inter-check context sharing (#1627). Build the read-only `ctx.priorFindings` a
@@ -442,40 +147,23 @@ export async function buildEditorialContext(seriesId, enabled, { providerOverrid
     needsContinuityBible ? getFactsLedger(seriesId).catch(() => null) : Promise.resolve(null),
   ]);
   const manuscript = sectionsCorpus(sections);
-  // The continuity-bible facts ledger (#1581) — the facts array, injected for the
-  // timeline / canon-contradiction check and fingerprinted for its staleness.
-  const continuityBible = Array.isArray(bible?.facts) ? bible.facts : [];
-  // Storyboard shots for the visual.shot-continuity check (#1315) — projected off
-  // the gated `issues` fetch (empty unless a storyboard.shots/comicScript check is on).
-  const storyboardScenes = collectStoryboardScenes(issues);
-  const reverseOutline = Array.isArray(outline?.scenes) ? outline.scenes : [];
-  // The outline's plotline list (#1310) — injected separately from the scenes so a
-  // plotline-reading check (plot.structure-momentum) can reconcile dropped subplots
-  // against the author's tagged threads.
-  const reverseOutlinePlotlines = Array.isArray(outline?.plotlines) ? outline.plotlines : [];
-  const editorialArcs = projectEditorialArcs(editorial);
-  // Whether every analyzable issue has been analyzed and is fresh — gates the
-  // pov.justified "absent from detected arcs" finding so a partially-analyzed
-  // series (canceled/early-stopped batch) doesn't flag a not-yet-analyzed POV
-  // holder as arc-less (#1295). Folded into the editorialArcs fingerprint below.
-  const editorialArcsComplete = editorialCoverageComplete(editorial);
-  // The comic content the comic checks read — derived from the already-loaded
-  // issues (no extra I/O). Parse each issue's comic script ONCE, then build all
-  // three per-token projections from the shared rows: lettering (#1313,
-  // caption/dialogue/SFX), pacing (#1314, + visual `description`), and layout
-  // (#1314, panel counts only) — so a description edit stales a pacing finding
-  // without staling a lettering one, and a text-only edit never stales a rhythm one.
-  const comicIssues = comicLetteringIssues(issues);
-  const comicScripts = projectComicLetteringContent(comicIssues);
-  const comicPacingContent = projectComicPacingContent(comicIssues);
-  const comicLayoutContent = projectComicLayoutContent(comicIssues);
-  // The per-issue PROSE-stage content the comic↔prose-sync check compares against the
-  // comic (#1589) — built off the same already-loaded issues, keyed by issue number.
-  const proseContent = proseStageIssues(issues);
-  // Resolve every source token once — each finding's fingerprint reads from this
-  // so the editor flags it `stale` when the content that check actually read (its
-  // declared `sources`) drifts (#1345, #1387).
-  const resolvedSources = resolveSources({ manuscript, canon, continuityBible, series, reverseOutline, reverseOutlinePlotlines, editorialArcs, editorialArcsComplete, storyboardScenes, comicScripts, comicPacingContent, comicLayoutContent, proseContent });
+  const {
+    continuityBible,
+    storyboardScenes,
+    reverseOutline,
+    reverseOutlinePlotlines,
+    editorialArcs,
+    editorialArcsComplete,
+    resolvedSources,
+  } = buildEditorialSourceProjection({
+    manuscript,
+    canon,
+    series,
+    issues,
+    outline,
+    editorial,
+    bible,
+  });
   const baseCtx = {
     seriesId,
     series,
@@ -929,114 +617,6 @@ export async function buildReverseOutlineGateContext(seriesId, { outline } = {})
     canon,
     reverseOutline: Array.isArray(resolved?.scenes) ? resolved.scenes : [],
     reverseOutlinePlotlines: Array.isArray(resolved?.plotlines) ? resolved.plotlines : [],
-  };
-}
-
-/**
- * Read the manuscript review and annotate each editorial-check finding with a
- * `stale` flag (#1345): true when the content the check analyzed has changed
- * since the finding was seeded. Mirrors `editorialAnalysis.isSnapshotStale` —
- * recompute the current source hash and compare against the one stamped on the
- * finding. Completeness-pass comments have no registry `checkId`, so their
- * manuscript-only hash is evaluated directly. Findings without a
- * `sourceContentHash` (older peers / legacy records), or whose named check is
- * no longer registered, are left unannotated → the UI treats absent `stale` as
- * not-stale.
- *
- * Staleness is derived per-read (never stored), so it stays local to each
- * install's current content and never rides the synced review document.
- */
-export async function getReviewWithStaleness(seriesId) {
-  const review = await getReview(seriesId);
-  // Resolve checks against built-ins + the user's custom checks (#1346) so a
-  // custom-check finding still gets staleness annotation. Build the id→check map
-  // once (custom-check synthesis is not free) and look up per comment.
-  const settings = await getSettings();
-  const byId = new Map(getAllChecks(settings).map((c) => [c.id, c]));
-  const checkFor = (id) => byId.get(id) || null;
-  // Completeness comments deliberately have no checkId; a hash on one of those
-  // identifies the manuscript-only completeness pass. Named checks still need
-  // to exist in the active registry before they can be evaluated.
-  const evaluable = review.comments.filter((c) => c.checkId && c.sourceContentHash && checkFor(c.checkId));
-  const completeness = review.comments.filter((c) => !c.checkId && c.sourceContentHash);
-  if (!evaluable.length && !completeness.length) return review;
-  if (!evaluable.length) {
-    // Completeness-only reviews need one manuscript read and none of the canon,
-    // series, outline, or registry-check source plumbing below.
-    const sections = await collectManuscriptSections(seriesId);
-    const current = manuscriptSourceHash(sectionsCorpus(sections));
-    return {
-      ...review,
-      comments: review.comments.map((c) => (!c.checkId && c.sourceContentHash
-        ? { ...c, stale: c.sourceContentHash !== current }
-        : c)),
-    };
-  }
-  // Re-fingerprint each finding against its check's EFFECTIVE sources — own plus
-  // declared-dependency sources (#1627) — exactly as the seed path stamped it, so a
-  // dependency-consuming finding's hash is computed against the same content on read
-  // as on seed. With no deps this is the check's own sources, unchanged.
-  const sourcesFor = (id) => effectiveCheckSources(checkFor(id), byId);
-  // Only pay the manuscript-collection I/O when an evaluable check declares it as
-  // a source (mirrors the run path's gate, now source-derived rather than the bare
-  // needsManuscript flag so it stays correct as the source vocabulary grows).
-  const needsManuscript = completeness.length > 0
-    || evaluable.some((c) => sourcesFor(c.checkId).includes('manuscript'));
-  const needsReverseOutline = evaluable.some((c) => {
-    const sources = sourcesFor(c.checkId);
-    return sources.includes('reverseOutline') || sources.includes('reverseOutline.plotlines');
-  });
-  const needsEditorialArcs = evaluable.some((c) => sourcesFor(c.checkId).includes('editorialArcs'));
-  // Continuity-bible ledger re-fingerprint, gated like the others (#1581) — must
-  // mirror the run path's fetch so a timeline/canon finding's hash is computed
-  // against the SAME ledger content on read as on seed (else it always reads stale).
-  const needsContinuityBible = evaluable.some((c) => sourcesFor(c.checkId).includes('continuityBible'));
-  // Issues are fetched here only when a storyboard-shots (#1315) OR comic-script
-  // (#1313 lettering / #1314 pacing) finding needs re-fingerprinting — all derive
-  // from the issue records, so a single gated fetch serves them. Mirrors the other
-  // per-source I/O gates.
-  const needsStoryboards = evaluable.some((c) => sourcesFor(c.checkId).includes('storyboard.shots'));
-  const needsComicScript = evaluable.some((c) => {
-    const s = sourcesFor(c.checkId);
-    return s.includes('comicScript') || s.includes('comicScript.pacing') || s.includes('comicScript.layout');
-  });
-  // The prose source (#1589) reads the per-issue prose stage off the same `issues`
-  // fetch, so it gates the issues load too — a prose-only check still loads issues.
-  const needsProse = evaluable.some((c) => sourcesFor(c.checkId).includes('prose'));
-  const needsIssues = needsStoryboards || needsComicScript || needsProse;
-  const series = await getSeries(seriesId);
-  const [sections, canon, outline, editorial, issues, bible] = await Promise.all([
-    needsManuscript ? collectManuscriptSections(seriesId) : Promise.resolve([]),
-    getSeriesCanon(series),
-    needsReverseOutline ? getReverseOutline(seriesId).catch(() => null) : Promise.resolve(null),
-    // Reuse the already-loaded series.
-    needsEditorialArcs ? getSeriesEditorial(seriesId, { series }).catch(() => null) : Promise.resolve(null),
-    needsIssues ? listIssuesForSeries(seriesId).catch(() => []) : Promise.resolve([]),
-    needsContinuityBible ? getFactsLedger(seriesId).catch(() => null) : Promise.resolve(null),
-  ]);
-  const continuityBible = Array.isArray(bible?.facts) ? bible.facts : [];
-  const reverseOutline = Array.isArray(outline?.scenes) ? outline.scenes : [];
-  const reverseOutlinePlotlines = Array.isArray(outline?.plotlines) ? outline.plotlines : [];
-  const editorialArcs = projectEditorialArcs(editorial);
-  const editorialArcsComplete = editorialCoverageComplete(editorial);
-  const storyboardScenes = collectStoryboardScenes(issues);
-  const comicIssues = comicLetteringIssues(issues);
-  const comicScripts = projectComicLetteringContent(comicIssues);
-  const comicPacingContent = projectComicPacingContent(comicIssues);
-  const comicLayoutContent = projectComicLayoutContent(comicIssues);
-  const proseContent = proseStageIssues(issues);
-  const resolvedSources = resolveSources({ manuscript: sectionsCorpus(sections), canon, continuityBible, series, reverseOutline, reverseOutlinePlotlines, editorialArcs, editorialArcsComplete, storyboardScenes, comicScripts, comicPacingContent, comicLayoutContent, proseContent });
-  return {
-    ...review,
-    comments: review.comments.map((c) => {
-      if (!c.checkId && c.sourceContentHash) {
-        return { ...c, stale: c.sourceContentHash !== manuscriptSourceHash(resolvedSources.manuscript) };
-      }
-      const check = c.checkId && c.sourceContentHash ? checkFor(c.checkId) : null;
-      if (!check) return c;
-      const current = fingerprintForCheck(check, resolvedSources, byId);
-      return { ...c, stale: c.sourceContentHash !== current };
-    }),
   };
 }
