+++ server/services/pipeline/issueStages.js
@@ -63,11 +63,11 @@ export function updateStagesWithLatest(seriesId, updates = [], { snapshotPrior =
     }
   }
 
-  let frontCoverChanged = false;
   return queueSeriesIssuesWrite(seriesId, async () => {
     const state = await readState();
     const results = [];
     let changed = false;
+    let frontCoverChanged = false;
     // Index once: `state.issues` is the INSTALL-wide list, and a bulk caller can
     // pass one update per (issue, stage) — a per-update findIndex would make the
     // write O(updates × issues in the install). Rebuilt in step with the
@@ -103,8 +103,8 @@ export function updateStagesWithLatest(seriesId, updates = [], { snapshotPrior =
       await saveIssuesNow(state.issues.filter((i) => i.seriesId === seriesId));
       emitRecordUpdated('series', seriesId);
     }
-    return results;
-  }).then(async (results) => {
+    return { results, frontCoverChanged };
+  }).then(async ({ results, frontCoverChanged }) => {
     if (frontCoverChanged) {
       // A manual clear/replacement does not pass through the media filename
       // hook, so it previously left series.coverImage pointing at stale art.
@@ -172,7 +172,6 @@ export function updateStageWithLatest(issueId, stageId, computeFn, { snapshotPri
   // see CLAUDE.md "single tail per shared file". seriesId is immutable, so read
   // it outside the lock to pick the queue, then re-read the issue INSIDE the
   // lock for the freshest stage.
-  let frontCoverChanged = false;
   const work = async () => {
     const cur = await store().loadOne(issueId);
     if (!cur) throw makeErr(`Issue not found: ${issueId}`, ERR_NOT_FOUND);
@@ -184,16 +183,17 @@ export function updateStageWithLatest(issueId, stageId, computeFn, { snapshotPri
     // page). Skip the disk write + emitRecordUpdated so it doesn't trigger
     // a re-export storm in share subscriptions for late no-op events.
     if (isPlainObject(patch) && Object.keys(patch).length === 0) {
-      return { issue: cur, stage: currentStage };
+      return {
+        result: { issue: cur, stage: currentStage },
+        frontCoverChanged: false,
+      };
     }
     // Snapshot the prior `{ runId, input, output }` into runHistory when this
     // patch carries a fresh lastRunId (i.e. a generate just replaced prior
     // content). Computed BEFORE the spread so it reads pre-merge state.
     const next = mergeStagePatch(currentStage, stageId, patch, { snapshotPrior });
-    if (stageId === 'comicPages'
-      && pickRenderedFilename(currentStage?.cover) !== pickRenderedFilename(next?.cover)) {
-      frontCoverChanged = true;
-    }
+    const frontCoverChanged = stageId === 'comicPages'
+      && pickRenderedFilename(currentStage?.cover) !== pickRenderedFilename(next?.cover);
     const mergedIssue = sanitizeIssue({
       ...cur,
       stages: { ...cur.stages, [stageId]: next },
@@ -201,10 +201,13 @@ export function updateStageWithLatest(issueId, stageId, computeFn, { snapshotPri
     });
     await saveIssueNow(mergedIssue);
     emitRecordUpdated('series', mergedIssue.seriesId);
-    return { issue: mergedIssue, stage: mergedIssue.stages[stageId] };
+    return {
+      result: { issue: mergedIssue, stage: mergedIssue.stages[stageId] },
+      frontCoverChanged,
+    };
   };
   return getIssue(issueId, { includeDeleted: true }).then((existing) =>
-    queueSeriesIssuesWrite(existing.seriesId, work).then(async (result) => {
+    queueSeriesIssuesWrite(existing.seriesId, work).then(async ({ result, frontCoverChanged }) => {
       if (frontCoverChanged) {
         const { refreshSeriesCoverImage } = await import('./seriesCoverImage.js');
         await refreshSeriesCoverImage(result.issue.seriesId).catch(() => {});
